requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
class-registry==2.1.2
cryptography==45.0.4
decouple==0.0.7
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
django-oauth-toolkit==3.0.1
django-ratelimit==4.1.0
django-restframework==0.0.1
djangorestframework==3.16.0
djangorestframework-oauth==1.1.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filters==1.3.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
mysql-connector-python==9.3.0
mysqlclient==2.2.7
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
timedelta==2020.12.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
